Feature/middleware validation

+++ routes/fighterRoutes.js
@@ -8,56 +8,48 @@ import {
 
 const router = Router();
 
-// TODO: Implement route controllers for fighter
+// TODO: Implement route controllers for fighter DONE!
 
 // GET route for retrieving all fighters
 router.get("/", async (req, res, next) => {
-  try {
+  if (!req.error_message) {
     const fighters = await fighterService.getAllFighters();
     res.json(fighters);
-  } catch (err) {
-    next(err);
-  }
+  } 
 });
 
 // POST route for creating a new fighter
-router.post("/", createFighterValid, async (req, res, next) => {
-  try {
+router.post("/", createFighterValid, responseMiddleware, async (req, res, next) => {
+  if (!req.error_message) {
     const fighter = await fighterService.createFighter(req.body);
     res.json(fighter);
-  } catch (err) {
-    next(err);
-  }
+  } 
 });
 
 // GET route for retrieving a fighter by ID
-router.get("/:id", async (req, res, next) => {
-  try {
+router.get("/:id", updateFighterValid, responseMiddleware, async (req, res, next) => {
+  if (!req.error_message) {
     const fighter = await fighterService.getFighterById(req.params.id);
+    //Turn first letter of thefighters name to uppercase
+    fighter.name = fighter.name.charAt(0).toUpperCase() + fighter.name.slice(1);
     res.json(fighter);
-  } catch (err) {
-    next(err);
-  }
+  } 
 });
 
 // PATCH route for updating a fighter
-router.patch("/:id", updateFighterValid, async (req, res, next) => {
-  try {
+router.patch("/:id", updateFighterValid, responseMiddleware, async (req, res, next) => {
+  if (!req.error_message) {
     const fighter = await fighterService.updateFighter(req.params.id, req.body);
     res.json(fighter);
-  } catch (err) {
-    next(err);
-  }
+  } 
 });
 
 // DELETE route for deleting a fighter
-router.delete("/:id", async (req, res, next) => {
-  try {
+router.delete("/:id", updateFighterValid, responseMiddleware, async (req, res, next) => {
+  if (!req.error_message) {
     await fighterService.deleteFighter(req.params.id);
     res.json({ message: "Fighter deleted successfully" });
-  } catch (err) {
-    next(err);
-  }
+  } 
 });
